+++ tcex/api/tc/ti_transform/__init__.py
@@ -1,6 +1,17 @@
 """TcEx Framework Module"""
 
 # first-party
+from tcex.api.tc.ti_transform.ti_predefined_functions import (
+    ProcessingFunctions,
+    transform_builder_to_model,
+)
 from tcex.api.tc.ti_transform.ti_transform import TiTransform, TiTransforms
+from tcex.api.tc.ti_transform.transform_abc import TransformException
 
-__all__ = ['TiTransform', 'TiTransforms']
+__all__ = [
+    'ProcessingFunctions',
+    'TiTransform',
+    'TiTransforms',
+    'TransformException',
+    'transform_builder_to_model',
+]