+++ tcex/api/tc/ti_transform/transform_abc.py
@@ -11,7 +11,6 @@ from typing import Any, cast
 
 # third-party
 import jmespath
-from jmespath import functions
 
 # first-party
 from tcex.api.tc.ti_transform import ti_predefined_functions
@@ -30,29 +29,30 @@ from tcex.api.tc.ti_transform.model.transform_model import (
     PredefinedFunctionModel,
 )
 from tcex.logger.trace_logger import TraceLogger
+from tcex.pleb.jmespath_custom import TcFunctions
 from tcex.util import Util
 
 # get tcex logger
 _logger: TraceLogger = logging.getLogger(__name__.split('.', maxsplit=1)[0])  # type: ignore
 
 
-class TcFunctions(functions.Functions):
-    """ThreatConnect custom jmespath functions."""
+class TransformException(Exception):
+    """Base exception for transform errors."""
 
-    @functions.signature({'types': ['array']}, {'types': ['string']})
-    def _func_null_leaf(self, arr: list, search: str) -> list:
-        """Return value in array even if they are null."""
-        return [a.get(search) for a in arr]
+    def __init__(self, field: str, cause: Exception, context: dict | None, *args) -> None:
+        """."""
+        super().__init__(*args)
+        self.field = field
+        self.cause = cause
+        self.context = context
 
-    @functions.signature({'types': ['array']}, {'types': ['string']})
-    def _func_delete(self, arr: list, search: str) -> list:
-        """Return array after popping value at address out."""
-        for a in arr:
-            try:
-                a.pop(search)
-            except Exception:
-                _logger.debug('Skipping delete, field not found.')
-        return arr
+    def __str__(self) -> str:
+        """."""
+        return f'Error transforming {self.field}: {self.cause}'
+
+
+class NoValidTransformException(Exception):
+    """Exception for when no valid transform is found."""
 
 
 class TransformsABC(ABC):
@@ -62,6 +62,7 @@ class TransformsABC(ABC):
         self,
         ti_dicts: list[dict],
         transforms: list[GroupTransformModel | IndicatorTransformModel],
+        raise_exceptions: bool = False,
     ):
         """Initialize instance properties."""
         self.ti_dicts = ti_dicts
@@ -69,6 +70,7 @@ class TransformsABC(ABC):
 
         # properties
         self.log = _logger
+        self.raise_exceptions = raise_exceptions
         self.transformed_collection: list[TransformABC] = []
 
         # validate transforms
@@ -78,7 +80,7 @@ class TransformsABC(ABC):
         """Validate the transform model."""
         if len(self.transforms) > 1:
             for transform in self.transforms:
-                if transform.applies is None:
+                if not callable(transform.applies):
                     raise ValueError(
                         'If more than one transform is provided, each '
                         'provided transform must provide an apply field.',
@@ -183,9 +185,14 @@ class TransformABC(ABC):
 
     def _process_associated_group(self, associations: list[AssociatedGroupTransform]):
         """Process Attribute data"""
-        for association in associations or []:
-            for value in filter(bool, self._process_metadata_transform_model(association.value)):
-                self.add_associated_group(value)  # type: ignore
+        for i, association in enumerate(associations or [], 1):
+            try:
+                for value in filter(
+                    bool, self._process_metadata_transform_model(association.value)
+                ):
+                    self.add_associated_group(value)  # type: ignore
+            except Exception as e:
+                raise TransformException(f'Associated Group [{i}]', e, context=association.dict())
 
     def _process_metadata_transform_model(
         self, value: bool | MetadataTransformModel | str | None, expected_length: int | None = None
@@ -388,64 +395,84 @@ class TransformABC(ABC):
 
     def _process_metadata(self, key: str, metadata: MetadataTransformModel | None):
         """Process standard metadata fields."""
-        value = self._transform_value(metadata)
-        if value is not None:
-            self.add_metadata(key, value)
+        try:
+            value = self._transform_value(metadata)
+            if value is not None:
+                self.add_metadata(key, value)
+        except Exception as e:
+            raise TransformException(key, e, metadata.dict() if metadata else None)
 
     def _process_metadata_datetime(self, key: str, metadata: DatetimeTransformModel | None):
         """Process metadata fields that should be a TC datetime."""
-        if metadata is not None and metadata.path is not None:
-            value = self._path_search(metadata.path)
-            if value is not None:
-                self.add_metadata(
-                    key, self.util.any_to_datetime(value).strftime('%Y-%m-%dT%H:%M:%SZ')
-                )
+        try:
+            if metadata is not None and metadata.path is not None:
+                value = self._path_search(metadata.path)
+                if value is not None:
+                    self.add_metadata(
+                        key, self.util.any_to_datetime(value).strftime('%Y-%m-%dT%H:%M:%SZ')
+                    )
+        except Exception as e:
+            raise TransformException(key, e, context=metadata.dict() if metadata else None)
 
     def _process_security_labels(self, labels: list[SecurityLabelTransformModel]):
         """Process Tag data"""
-        for label in labels or []:
-            names = self._process_metadata_transform_model(label.value)
-            if not names:
-                # self.log.info(f'No values found for security label transform {label.dict()}')
-                continue
-
-            descriptions = self._process_metadata_transform_model(
-                label.description, expected_length=len(names)
-            )
-            colors = self._process_metadata_transform_model(label.color, expected_length=len(names))
-
-            param_keys = ['color', 'description', 'name']
-            params = [dict(zip(param_keys, p)) for p in zip(colors, descriptions, names)]
-
-            for kwargs in params:
-                kwargs = self.util.remove_none(kwargs)
-                if 'name' not in kwargs:
+        for i, label in enumerate(labels or [], 1):
+            try:
+                names = self._process_metadata_transform_model(label.value)
+                if not names:
+                    # self.log.info(f'No values found for security label transform {label.dict()}')
                     continue
-                # strip out None params so that required params are enforced and optional
-                # params with default values are respected.
-                self.add_security_label(**kwargs)
+
+                descriptions = self._process_metadata_transform_model(
+                    label.description, expected_length=len(names)
+                )
+                colors = self._process_metadata_transform_model(
+                    label.color, expected_length=len(names)
+                )
+
+                param_keys = ['color', 'description', 'name']
+                params = [dict(zip(param_keys, p)) for p in zip(colors, descriptions, names)]
+
+                for kwargs in params:
+                    kwargs = self.util.remove_none(kwargs)
+                    if 'name' not in kwargs:
+                        continue
+                    # strip out None params so that required params are enforced and optional
+                    # params with default values are respected.
+                    self.add_security_label(**kwargs)
+            except Exception as e:
+                raise TransformException(f'Security Labels [{i}]', e, context=label.dict())
 
     def _process_tags(self, tags: list[TagTransformModel]):
         """Process Tag data"""
-        for tag in tags or []:
-            for value in filter(bool, self._process_metadata_transform_model(tag.value)):
-                self.add_tag(name=value)  # type: ignore
+        for i, tag in enumerate(tags or [], 1):
+            try:
+                for value in filter(bool, self._process_metadata_transform_model(tag.value)):
+                    self.add_tag(name=value)  # type: ignore
+            except Exception as e:
+                raise TransformException(f'Tags [{i}]', e, context=tag.dict())
 
     def _process_rating(self, metadata: MetadataTransformModel | None):
         """Process standard metadata fields."""
-        self.add_rating(self._transform_value(metadata))
+        try:
+            self.add_rating(self._transform_value(metadata))
+        except Exception as e:
+            raise TransformException('Rating', e, context=metadata.dict() if metadata else None)
 
     def _process_type(self):
         """Process standard metadata fields."""
-        _type = self._transform_value(self.transform.type)
-        if _type is not None:
-            self.transformed_item['type'] = _type
-        else:
-            self.log.error(
-                'feature=ti-transform, action=process-type, error=invalid=type, '
-                f'path={self.transform.type.path}, value={_type}'
-            )
-            raise RuntimeError('Invalid type')
+        try:
+            _type = self._transform_value(self.transform.type)
+            if _type is not None:
+                self.transformed_item['type'] = _type
+            else:
+                self.log.error(
+                    'feature=ti-transform, action=process-type, error=invalid=type, '
+                    f'path={self.transform.type.path}, value={_type}'
+                )
+                raise RuntimeError('Invalid type')
+        except Exception as e:
+            raise TransformException('Type', e, context=self.transform.type.dict())
 
     def _select_transform(self):
         """Select the correct transform based on the "applies" field."""
@@ -454,7 +481,7 @@ class TransformABC(ABC):
                 self.transform = transform
                 break
         else:
-            raise RuntimeError('No transform found for TI data')
+            raise NoValidTransformException('No transform found for TI data')
 
     def _transform_value(self, metadata: MetadataTransformModel | None) -> str | None:
         """Pass value to series transforms."""
@@ -557,7 +584,7 @@ class TransformABC(ABC):
 
         # return default if value of None is returned from Path
         # IMPORTANT: a None value passed to the transform may cause a failure (lambda x: x.lower())
-        if value in [None, []]:
+        if value in [None, [], '']:
             default = _default()
             # self.log.trace(
             #     f'feature=transform, action=transform-values, metadata-path=None, value={default}'