+++ tcex/app/service/webhook_trigger_service.py
@@ -1,4 +1,5 @@
 """TcEx Framework Module"""
+
 # standard library
 import base64
 import json
@@ -314,12 +315,14 @@ class WebhookTriggerService(CommonServiceTrigger):
             try:
                 # call callback method
                 # pylint: disable=not-callable
-                callback_response: dict | None = self.webhook_marshall_event_callback(
-                    body=body,
-                    headers=message.get('headers'),
-                    request_key=request_key,
-                    status_code=message.get('statusCode'),
-                    trigger_id=message.get('triggerId'),
+                callback_response: dict | None = (
+                    self.webhook_marshall_event_callback(  # type: ignore
+                        body=body,
+                        headers=message.get('headers'),
+                        request_key=request_key,
+                        status_code=message.get('statusCode'),
+                        trigger_id=message.get('triggerId'),
+                    )
                 )
                 if isinstance(callback_response, dict):
                     response.update(callback_response)